+++ src/trading_strategies/strategy/option_strategy/naked_call.py
@@ -1,4 +1,4 @@
-from src.trading_strategies.financial_asset.option import PutOption, Option, CallOption
+from src.trading_strategies.financial_asset.option import Option, CallOption
 from src.trading_strategies.financial_asset.price import Price
 from src.trading_strategies.financial_asset.symbol import Symbol
 from src.trading_strategies.option_pricing import bsm_pricing, implied_date
@@ -6,7 +6,7 @@ from src.trading_strategies.strategy.option_strategy.option_strategy import Opti
 from src.trading_strategies.strategy.option_strategy.option_strike import calculate_strike, \
     roll_up_strike
 from src.trading_strategies.strategy.strategy_id import StrategyId
-from src.trading_strategies.transactions.position import Position
+from src.agent.transactions.position import Position
 from src.util.expiry_date import closest_expiration_date, nyse_calendar
 
 risk_free_rate = 0.03
