+++ src/trading_strategies/strategy/option_strategy/naked_put.py
@@ -5,7 +5,7 @@ from src.trading_strategies.option_pricing import bsm_pricing, implied_date
 from src.trading_strategies.strategy.option_strategy.option_strategy import OptionStrategy
 from src.trading_strategies.strategy.option_strategy.option_strike import calculate_strike, roll_down_strike
 from src.trading_strategies.strategy.strategy_id import StrategyId
-from src.trading_strategies.transactions.position import Position
+from src.agent.transactions.position import Position
 from src.util.expiry_date import closest_expiration_date, nyse_calendar
 
 risk_free_rate = 0.03
